+++ packages/formstr-app/src/components/Header/configs.tsx
@@ -13,7 +13,6 @@ export const HEADER_MENU_KEYS = {
   PUBLIC_FORMS: "PUBLIC_FORMS",
   USER: "USER",
   CREATE_FORMS: "CREATE_FORMS",
-  IMPORT_FORMS: "IMPORT_FORMS",
   HELP: "HELP",
   CONTACT_US: "CONTACT_US",
 };
@@ -48,11 +47,6 @@ export const HEADER_MENU = [
         Create Form
       </Button>
     ),
-  },
-  {
-    key: HEADER_MENU_KEYS.IMPORT_FORMS,
-    label: "Import",
-    icon: <ImportOutlined />
   },
   {
     key: HEADER_MENU_KEYS.PUBLIC_FORMS,
