+++ packages/formstr-app/src/components/Header/index.tsx
@@ -30,12 +30,6 @@ import { useState } from "react";
 import { useTemplateContext } from "../../provider/TemplateProvider";
 import ThemedUniversalModal from "../UniversalMarkdownModal";
 import { nip19 } from "nostr-tools";
-import GoogleFormsDeployer, {
-  FetchResult,
-} from "../GoogleFormsDeployer/index";
-import { makeTag } from "../../utils/utility";
-import { ROUTES } from "../../constants/routes";
-import { mapGoogleFormQuestionsToFieldsAndSections } from "../GoogleFormsDeployer/helper";
 
 const { Text, Paragraph } = Typography;
 
@@ -59,53 +53,12 @@ export const NostrHeader = () => {
     disableEncryption,
   } = useLocalForms();
   const [isFAQModalVisible, setIsFAQModalVisible] = useState(false);
-  const [isImportModalVisible, setIsImportModalVisible] = useState(false);
   const [showEncryptionModal, setShowEncryptionModal] = useState(false);
   const [encryptionLoading, setEncryptionLoading] = useState(false);
   const [selectedKey, setSelectedKey] = useState<string[]>([]);
   const { openTemplateModal } = useTemplateContext();
   const navigate = useNavigate();
 
-  const handleGoogleFetchRender = (result: FetchResult) => {
-    if (!result.success || !Array.isArray(result.data?.questions)) {
-      message.error("No form schema available to render.");
-      return;
-    }
-    const { fields, sections } = mapGoogleFormQuestionsToFieldsAndSections(
-      result.data!.questions,
-    );
-    if (!fields.length) {
-      message.warning("No supported questions found in this Google Form.");
-      return;
-    }
-    const formId = makeTag(6);
-    const spec = [
-      ["d", formId],
-      ["name", result.data?.title || "Imported Google Form"],
-      [
-        "settings",
-        JSON.stringify({
-          thankYouPage: true,
-          publicForm: true,
-          disallowAnonymous: false,
-          encryptForm: true,
-          viewKeyInUrl: true,
-          description: result.data?.description || "",
-          ...(sections.length > 0 ? { sections } : {}),
-        }),
-      ],
-      ...fields,
-    ];
-    navigate(ROUTES.CREATE_FORMS_NEW, {
-      state: {
-        spec,
-        id: formId,
-      },
-    });
-    setIsImportModalVisible(false);
-    setSelectedKey([]);
-    message.success("Google Form rendered in Form Builder.");
-  };
 
   const handleEnableEncryption = async () => {
     setEncryptionLoading(true);
@@ -321,11 +274,6 @@ export const NostrHeader = () => {
       openTemplateModal();
       return;
     }
-    if (e.key === HEADER_MENU_KEYS.IMPORT_FORMS) {
-      setIsImportModalVisible(true);
-      setSelectedKey([e.key]);
-      return;
-    }
     setSelectedKey([e.key]);
   };
 
@@ -443,17 +391,6 @@ export const NostrHeader = () => {
         filePath="/docs/faq.md"
         title="Frequently Asked Questions"
       />
-      <Modal
-        open={isImportModalVisible}
-        onCancel={() => setIsImportModalVisible(false)}
-        footer={null}
-        width={1000}
-      >
-        <GoogleFormsDeployer
-          onFetch={(json) => console.log(json)}
-          onRenderInBuilder={handleGoogleFetchRender}
-        />
-      </Modal>
       <Modal
         title={getEncryptionModalTitle()}
         open={showEncryptionModal}
